requirements.txt
PySide6>=6.6
Pillow>=10.0
mss>=9.0
pynput>=1.7
pyautogui>=0.9
